fix: generic component for bill section and bill list
